+++ Validator/FormatDateValidator.php
@@ -20,6 +20,8 @@ use function preg_match;
 use function sprintf;
 
 /**
+ * This validation based on rfc3339 {@see https://xml2rfc.tools.ietf.org/public/rfc/html/rfc3339.html#anchor14}
+ *
  * @author Viktor Linkin <adrenalinkin@gmail.com>
  */
 class FormatDateValidator implements SwaggerValidatorInterface